// server/src/modules/routes/core/domain/dtos/CreateRoute.dto.ts
import { RouteType } from '../shared/types/route-type.enum';
import { RouteStatus } from '../shared/types/route-status.enum';

export interface CreateRouteDTO {
  id?: string;
  code: string;
  name: string;
  shortName?: string;
  type: RouteType;
  status: RouteStatus;
  color: string;
  icon?: string;
  totalDistanceKm: number;
  estimatedDurationMin: number;
  polyline: string;
  boundingBox?: {
    ne: { lat: number; lng: number };
    sw: { lat: number; lng: number };
  };
}